- Removed some unnecessary data structures in Menu control (defined its own rect when Control already has one). - Structure now has an 'isFactory' field. - Fleshed out the Factory interface class a bit more. It's now generally functional. - SeedFactory and SeedSmelter now derive from Factory instead of Structure. - SeedSmelter is now processing raw resources. - Tinkered with Agridome and CHAP with their resource consumption. Will require better balancing to be playable. - StructureManager now implements a factory update loop. - Pulled menuFactoryProduction -- going to go with a full blown window interface instead. - Pulled unused/legacy typedef from GameState. - Added a changeDepth() function to GameState which is called whenever number keys are pressed (code readability). - Changed Resources structure to store resources as double's instead of int's. Helps to correct for rounding issues and makes more sense in terms of a simulation. - Played with the mine yields a bit... still not where I want them to be but better.
git-svn-id: https://svn.outpostuniverse.org:8443/svn/outposthd@22 bdf9137d-f3be-7946-88fe-7b0bbfdbf31a

+++ src/Resources.h
@@ -90,16 +90,16 @@ public:
 	 }
 
 	// Ore
-	int	commonMetalsOre;
-	int	commonMineralsOre;
-	int	rareMetalsOre;
-	int	rareMineralsOre;
+	double commonMetalsOre;
+	double commonMineralsOre;
+	double rareMetalsOre;
+	double rareMineralsOre;
 
 	// Refined Materials
-	int	commonMetals;
-	int	commonMinerals;
-	int	rareMetals;
-	int	rareMinerals;
+	double commonMetals;
+	double commonMinerals;
+	double rareMetals;
+	double rareMinerals;
 
 	// Other
 	int	energy;
